e_commerce/requirements.txt
asgiref==3.7.2
Django==4.2.2
django-cleanup==8.0.0
django-cors-headers==4.1.0
django-coupon-management==1.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
psycopg2-binary==2.9.6
PyJWT==2.7.0
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
